Special Rules and a bit more
Added the text for unique special rules in the mech lib.
Renamed the detachments to match the others in the main file.
Added in the Arcana Benefice to all the Archmagos, Magos and Techpriests including those on Abeyants.

Added ability to select our sub-faction for all units which require the ability to select them.

Added Flareshield, Rad Furnace, Machinator Array wargear descriptions