Moving stuff around and adding Gambit and Cyber stuff
Moved some Mech Wargear to the Wargear lib as it will likley appear elsewhere such as Knights and Custodes.
Moved Corpasant Stave to Weapons lib.
Added Gambit and Cybertheurgic Rite profiles to GST as these are required for clarity.
Added the missing categories such as Unstoppable and Rapid and Vehicle to GST to add to Mech units which require such things for Prime bonus checks.